Bootstrap 4, popper, fontawesome Formularios con múltiples respuestas por usuario Edición wysiwyg usando sommernote

+++ api/controllers/AdminController.js
@@ -27,16 +27,16 @@ module.exports = {
 
 	download: function (req, res) {
 
-		Recibidos.find({formid:req.formId,borrado:false}).exec(function(err,recibido) {
+		Recibidos.find({formid:req.formId,borrado:false}).exec(function(err,recibidos) {
 			if (err) {
 				return res.serverError(err);
 			}
 
 			res.set('Content-type','text/tab-separated-values; charset=utf-8');
-			recibido.forEach(function(item) {
+			recibidos.forEach(function(item) {
 				item.updatedAt = item.updatedAt.fecha_ymdhms_toString();
 			});
-			return res.view("admin/download.ejs",{layout:'',config:req.config,recibido:recibido});
+			return res.view("admin/download.ejs",{layout:'',config:req.config,recibidos:recibidos});
 		});
 	},
 
@@ -102,10 +102,11 @@ module.exports = {
 
 	edit: function (req, res) {
 
-		var titulo = req.param('titulo').trim();
-		var desde = req.param('desde').trim();
-		var hasta = req.param('hasta').trim();
-		var publico = req.param('publico')=='on';
+		let titulo = req.param('titulo').trim();
+		let desde = req.param('desde').trim();
+		let hasta = req.param('hasta').trim();
+		let publico = req.param('publico')=='on';
+		let multiple = req.param('multiple')=='on';
 
 		if (!titulo || !desde || !hasta || titulo==="" || desde==="" || hasta==="") {
 			return res.json(500,{message:"Debe especificar título, fecha desde y fecha hasta"});
@@ -119,7 +120,7 @@ module.exports = {
 		// paso las fechas a UTCString y lo corro algunas horas porque hace conversión de TimeZone
 		desde=desde + 'T05:00:00.000Z';
 		hasta=hasta + 'T05:00:00.000Z';
-		Config.update({formid:req.formId},{titulo:titulo,desde:desde,hasta:hasta,publico:publico}).exec(function(err,updated){
+		Config.update({formid:req.formId},{titulo:titulo,desde:desde,hasta:hasta,publico:publico,multiple:multiple}).exec(function(err,updated){
 			if (err) {
 				return res.json(500,{message:err.message});
 			}
@@ -161,17 +162,24 @@ module.exports = {
 		} else if (action == "editar") {
 			Modulos.findOne({formid:req.formId,orden:orden}).exec(function(err,m){
 				if (err) {
-						return callback(err);
+						return res.serverError(err);
 				}
 				if (typeof m === 'undefined') {
-					return callback(new Error("No se encuentra el componente #"+orden));
+					return res.serverError(new Error("No se encuentra el componente #"+orden));
 				}
 				if (typeof sails.controllers[m.modid] === 'undefined' || typeof sails.controllers[m.modid].modedit === 'undefined') {
 					return res.serverError(new Error("Los componentes de tipo "+m.modid+" no pueden ser editados"));
 				}
+				["nombre","etiqueta","ayuda"].forEach(campo => {
+					if (m[campo] == null) {
+						m[campo]="";
+					}
+				});
 				sails.controllers[m.modid].load({config:req.config,m:m},function(){
 					m.config = sails.controllers[m.modid].modedit();
-					return res.view("admin/modedit.ejs",{title:'Editar Componente',formId:req.formId,m:m});
+					var modulos = Object.keys(sails.controllers).filter(x => (x !== "form" && x !== "admin"));
+					var validadores = Array({id:"nop",desc:"Sin Validador"},{id:"correo",desc:"Correo electrónico"},{id:"novacio",desc:"No vacío"},{id:"numero",desc:"Es número"},{id:"edad",desc:"Edad"});
+					return res.view("admin/modedit.ejs",{title:'Editar Componente',formId:req.formId,m:m,session:req.session,modulos:modulos,validadores:validadores});
 				},function(err){
 					return res.serverError(err);
 				});
@@ -184,6 +192,9 @@ module.exports = {
 				}
 				if (!m) {
 					return res.serverError(new Error("Módulo no encontrado para borrar"));
+				}
+				if (m.modid == "grupo_cierra") {
+					return res.serverError(new Error("Los módulos grupo_cierra no se pueden borrar"));
 				}
 				Modulos.destroy({formid:req.formId,id:m.id}).exec(function(err){
 					if (err) {
@@ -211,45 +222,52 @@ module.exports = {
 			});
 
 		} else if (action == "crear") {
-			return res.redirect(sails.config.baseurl + 'form/modedit?formId='+req.formId);
-
-		} else if (action == "guardar") {
-			var nombre=req.param("nombre");
-			var etiqueta=req.param("etiqueta");
-			var texto1=req.param("texto1");
-			var texto2=req.param("texto2");
-			var ayuda=req.param("ayuda");
-			var validador=req.param("validador");
-			var opcional=req.param("opcional") === "on";
-			Modulos.update({formid:req.formId,orden:orden},{nombre:nombre,etiqueta:etiqueta,texto1:texto1,texto2:texto2,ayuda:ayuda,validador:validador,opcional:opcional}).exec(function(err,updated){
+			Modulos.findOne({formid:req.formId,orden:{'<':999}}).sort('orden DESC').exec(function(err,m){
 				if (err) {
 					return res.serverError(err);
 				}
-				return res.redirect(sails.config.baseurl + 'form/modedit?formId='+req.formId);
-			});
-
-		} else if (action == "guardar_siguiente") {
-			var nombre=req.param("nombre");
-			var etiqueta=req.param("etiqueta");
-			var texto1=req.param("texto1");
-			var texto2=req.param("texto2");
-			var ayuda=req.param("ayuda");
-			var validador=req.param("validador");
-			var opcional=req.param("opcional") === "on";
-			Modulos.update({formid:req.formId,orden:orden},{nombre:nombre,etiqueta:etiqueta,texto1:texto1,texto2:texto2,ayuda:ayuda,validador:validador,opcional:opcional}).exec(function(err,updated){
-				if (err) {
-					return res.serverError(err);
-				}
-				Modulos.findOne({formid:req.formId,orden:{'>':orden}}).sort('orden').exec(function(err,m){
+				let orden = (!m ? 1 : m.orden+1);
+				let modid = "input_text";
+				Modulos.create({formid:req.formId,modid:modid,orden:orden}).exec(function(err,created){
 					if (err) {
 						return res.serverError(err);
 					}
-					if (!m) {
-						// no hay siguiente
-						return res.redirect(sails.config.baseurl + 'form/modedit?formId='+req.formId);
-					}
-					return res.redirect(sails.config.baseurl + 'admin/modedit?formId='+req.formId+'&orden='+m.orden+'&action=editar');
+					return res.redirect(sails.config.baseurl + 'admin/modedit?formId='+req.formId+'&orden='+orden+'&action=editar');
 				});
+			});
+
+		} else if (action == "guardar" || action == "guardar_siguiente") {
+			let modid=req.param("modid");
+			let nombre=req.param("nombre");
+			let etiqueta=req.param("etiqueta");
+			let texto1=req.param("texto1");
+			let texto2=req.param("texto2");
+			let ayuda=req.param("ayuda");
+			let validador=req.param("validador");
+			let opcional=req.param("opcional") === "on";
+			if (!modid) {
+				return res.serverError(new Error("Parámetros incorrectos"));
+			}
+			let ordenOriginal = orden;
+			orden = (modid == "intro" ? 1 : (modid == "pie" ? 999 : orden));
+			Modulos.update({formid:req.formId,orden:ordenOriginal},{modid:modid,orden:orden,nombre:nombre,etiqueta:etiqueta,texto1:texto1,texto2:texto2,ayuda:ayuda,validador:validador,opcional:opcional}).exec(function(err,updated){
+				if (err) {
+					return res.serverError(err);
+				}
+				if (action == "guardar") {
+					return res.redirect(sails.config.baseurl + 'admin/modedit?formId='+req.formId+'&orden='+orden+'&action=editar');
+				} else {
+					Modulos.findOne({formid:req.formId,orden:{'>':orden}}).sort('orden').exec(function(err,m){
+						if (err) {
+							return res.serverError(err);
+						}
+						if (!m) {
+							// no hay siguiente
+							return res.redirect(sails.config.baseurl + 'form/modedit?formId='+req.formId);
+						}
+						return res.redirect(sails.config.baseurl + 'admin/modedit?formId='+req.formId+'&orden='+m.orden+'&action=editar');
+					});
+				}
 			});
 
 		} else {
